Added lists for New Zealand female and male first names (1,564 and 1,172 respectively). The intention is to use these names as easy to remember alias ids that are unique, so when discussing a particular case it can be refered to by a human name (easier than numbers/codes). These two lists have been placed in a folder called meta, which should be used to store data that has been prepared in advance for input into other data generation scripts (such as a name id assigning script). How they are to be assigned is yet to be decided, however if there is a concern there is not enough (since there are only 2,736), aspects like DHB could also be concatenated - e.g. Jim-AK or Jim-BP (Jim from Auckland and Jim from Bay of Plenty).
The two lists were derived from the 'Baby Name Popularity Over Time' list from the NZ Department of Internal Affairs, which is a list of common New Zealand names registered at birth dating back to the 1900s to now. The list was separated into two lists by gender, then a distinct and shuffle operation performed. Lastly the license for using this list is a Creative Commons 4.0.