Added cleaned up version of the first 262 cases. The extra fields 'transmission' and 'cluster' are included.
The 'transmission' field takes one of three values:
a) overseas - exposed via international travel
b) linked - exposure linked to a confirmed/probable case or known cluster
c) community - source of exposure unknown and no travel outside of New Zealand, therefore it is community transmission
If there is no value/null, this is simply because the information was not recorded - in reality most of these cases will be linked cases because of low community transmission in NZ and the absence of any flight details. Later records may provide new information to help select a value for transmission for these null values.

The 'cluster' field indicates if the case is linked to any known clusters or sources of travel such as cruise ships. If a case's transmission value is 'linked' however the cluster value is null, this means the case was linked to a known individual (as opposed to a cluster).

Lastly, moved and renamed the cruise.csv to also be meta data of the first 262 cases.