More changes made dirty data surfacing in record linking with meta data, namely the several different spellings of Hawke's Bay and use of different apostrphe characters, as well as removing the hyphen betwee Nelson-Marlborough.
Also changed was "Unknown" -> "Not Recorded", this is a better descriptor, because information may later be known through linking/enrichment, it just simply isn't recorded at the time. Secondly "Unknown" also is synonymously used with Community Transmission, when a case is not connected to any known cases, so it is best to avoid overloading the terminology.

Case 51, "Travel related, however international flights outside of infectious period. Domestic flights: Auckland to Nelson on 16 March – flight 5065" -> Changed this from 'Local Community' to 'Overseas' transmission. While the comment suggests overseas transmission is not possible due to the infectious period, this later makes it not possible to link an extremely closely matched case in later records. Therefore it appears later records conclude that overseas transmission was the cause.